Extract make_key_pair into a reusable function
