+++ crates/bitwarden/src/crypto/mod.rs
@@ -44,6 +44,8 @@ pub(crate) use master_key::{HashPurpose, MasterKey};
 mod user_key;
 #[cfg(feature = "internal")]
 pub(crate) use user_key::UserKey;
+#[cfg(feature = "internal")]
+mod rsa;
 
 #[cfg(feature = "internal")]
 mod fingerprint;
